Make day-02 more functional with fold
Also make less generic by always initializing from horizontal = depth = aim = 0 position. Folding a tuple with match is a nice pattern for this kind of imperative series of instructions. Could fold with the Submarine struct directly, but it would add a lot of visual noise that might obscure the logic.